add some sleep to wait data fill in audio packet queue
The behavior of cooperation of FFmpeg and AudioQueue should be revised.
We should avoid to use sleep() to get the correct time to start
AudioQueue, otherwise the sleep time may differ bet. different hardware.
We may calculate packet count, and start AudioQueue when packet count
is bigger than a fixed number (at least 3. This is the value of
AUDIO_BUFFER_QUANTITY).